Move logic to ui_options.cpp
